Remove legacy view
