+++ src/views/results/explain/icons.ts
@@ -1,0 +1,48 @@
+
+/**
+ * Icon labels as defined by the API, along with the name of the icon to display.
+ * Not surprisingly, the reference link does not provide a complete list of icons.
+ * TODO: Add missing icons
+ * @see https://www.ibm.com/docs/en/i/7.5?topic=ssw_ibm_i_75/apis/qqqvexpl.html#icon_labels
+ * @see https://code.visualstudio.com/api/references/icons-in-labels
+ */
+const icons = {
+  "Bitmap Merge":                           `merge`,
+  "Cache":                                  ``,
+  "Cache Probe":                            ``,
+  "Delete":                                 `trash`,
+  "Distinct":                               `list-flat`,
+  "Dynamic Bitmap":                         `symbol-misc`,
+  "Encoded Vector Index":                   `symbol-reference`,
+  "Encoded Vector Index, Parallel":         `symbol-reference`,
+  "Final Select":                           `selection`,
+  "Hash Grouping":                          `group-by-ref-type`,
+  "Hash Join":                              `add`,
+  "Hash Scan":                              `search`,
+  "Index Grouping":                         `group-by-ref-type`,
+  "Index Scan - Key Positioning":           `key`,
+  "Index Scan - Key Positioning, Parallel": `key`,
+  "Index Scan - Key Selection":             `key`,
+  "Index Scan - Key Selection, Parallel":   `key`,
+  "Insert":                                 `insert`,
+  "Nested Loop Join":                       `add`,
+  "Select":                                 `selection`,
+  "Skip Sequential Table Scan":             `list-unordered`,
+  "Skip Sequential Table Scan, Parallel":   `list-unordered`,
+  "Sort":                                   `sort-precedence`,
+  "Sorted List Scan":                       `list-ordered`,
+  "Subquery Merge":                         `merge`,
+  "Table Probe":                            `list-selection`,
+  "Table Scan":                             `search`,
+  "Table Scan, Parallel":                   `search`,
+  "Temporary Distinct Hash Table":          `new-file`,
+  "Temporary Hash Table":                   `new-file`,
+  "Temporary Index":                        `new-file`,
+  "Temporary Sorted List":                  `list-ordered`,
+  "Temporary Table":                        `new-file`,
+  "Union Merge":                            `merge`,
+  "User Defined Table Function":            `symbol-function`,
+  "Unknown":                                `question`,
+  "Update":                                 `replace`,
+  "VALUES LIST":                            `list-flat`,
+}
